requirements.txt
requests>=2.31.0
PyAutoGUI>=0.9.54
comtypes>=1.3.0
pycaw>=20230407
python-dotenv>=1.0.1
fastapi>=0.109.2
starlette>=0.36.3
uvicorn>=0.27.1